Refactor tests and move things around a bit.
