+++ src/value/ser.rs
@@ -11,8 +11,8 @@ use std::collections::BTreeMap;
 use crate::error::Error;
 use serde::{self, Serialize};
 
+use crate::tags::{set_tag, CBOR_NEWTYPE_NAME};
 use crate::value::Value;
-use crate::{set_tag, CBOR_NEWTYPE_NAME};
 
 impl serde::Serialize for Value {
     #[inline]
